Restructure package to match 1.0 proposal (#207)
## Motivation

As part of the roadmap for 1.0 (see #206), we need to restructure the
package to reflect the accepted proposal (#205), which will reduce the
API surface significantly.

## Modifications

- Bump `swift-tools-version` to 6.1
- Rename `OTel` module to `OTelCore`
- Remove `OTelCore` and `OTLPGRPC` library products
- Add new, empty `OTel` target and library product
- Define new `OTLPHTTP` and `OTLPGRPC` traits (enabled by default)
- Add new, empty `OTLPHTTP` internal target
- Remove old API use from examples and replace with placeholder

## Result

Package now has the proposed structure, which will unblock future work
from the 1.0 roadmap.

## Notes

To keep this PR targeted and easy to review, it does not update the
visibility of the symbols in `OTelCore`, which will follow in a
subsequent PR.