Mark all previously public API as package visibile (#209)
## Motivation

As part of the roadmap for 1.0 (see #206), we need to reduce the public
API surface to reflect the accepted proposal (#205).

## Modifications

- Replace all `public` and `@_spi(...) public` access with `package`
access
- Update Makefile to generate Swift from proto with package visibility

## Result

There is no longer any `public` symbols in the repo:

```console
% git grep public Sources/ | wc -l
0

% git grep package Sources/ | wc -l
1032
```

This gives us a ground-zero from which to add new public API surface to
the new `OTel` module.

## Notes

To keep this PR targeted and easy to review, this is a very mechanical
change. _Everything_ that was `public` is now `package`. If we want to
discuss further changing the visibility, e.g. to `internal` for some
symbols, this is deferred for now.