Add stubbed public API surface for configuration, bootstrap, and backends (#210)
## Motivation

As part of the roadmap for 1.0 (see #206), we need to add the public API
surface from the proposal (#205).

## Modifications

- Add `OTel.Configuration` to `OTelCore` (it will be needed by the other
modules).
- Add stub for `OTel.bootstrap`.
- Add stubs `OTel.makeLoggingBackend`, `OTel.makeMetricsBackend`, and
`OTel.makeTracingBackend`.
- Add public API tests.

## Result

Public API exists to match the proposal. The API currently is only
stubbed and throws a `NotImplemented` error, but will unblock more work,
which can start to be parallelised.

## Notes

1. To keep this PR targeted and easy to review, it only adds the symbols
agreed in the proposal. One notable thing it does _not_ include is API
documentation, which I have in a separate branch. This is because that's
likely to attract much more commentary during review.

2. There are values in the configuration that this package does not
support but were added because they were (a) part of the API proposal,
and (b) part of the OTel spec. I think we should consider removing them,
but I have added this as an open question on the roadmap to not block
this PR.