Add shims to initialize internal types from new configuration type (#211)
## Motivation

As part of the roadmap for 1.0 (see #206), we have introduced a new
centralized configuration API as part of the API proposal (#205). The
plan is to continue to make use of all the (now-)internal types to
implement the API, and many of these had some level of configuration,
either with a different configuration type (which will also now be
internal), or by initializer parameters, which are no longer reachable
by API. We need a way to construct the hierarchy of internal types from
the new, user-provided `OTel.Configuration` type.

## Modifications

-  Add `init(_: OTel.Configuration)` shims for existing types.

## Result

Can now construct all the types needed to create and bootstrap backends
from an `OTel.Configuration` value.

## Notes

To keep this PR targeted and easy to review, the following are out of
scope, and will be provided in future PRs:

1. Supporting a wider range of configuration. Specifically, the new
`OTel.Configuration` type allows for expressing more configuration than
the previous API supported. This PR only wires up what was previously
supported.

2. Removing the internal existing initializer or configuration APIs. A
future PR might remove any obsolete configuration APIs.