Add internal `Configuration.applyEnvironmentOverrides` helpers (#212)
## Motivation

As part of the roadmap for 1.0 (see #206), we have introduced a new
centralized configuration API as part of the API proposal (#205). The
implementation also needs to support applying overrides from environment
variables as described in the OTel spec.

## Modifications

- Add internal `Configuration.applyEnvironmentOverrides` helpers

## Result

Internal helper functions exist for overriding config using environment
variables, which can be used to implement the backend APIs.

## Notes

There's a note in the roadmap to discuss how we would like to handle
unsupported configuration values that are detected in the environment
variable. Right now they are handled as follows:

1. Unrecognised values revert to default. This is in line with the OTel
spec, although I have added an item to our roadmap to discuss whether we
want a strict mode. They should also be logged with the "internal
diagnostic" logger, which we don't have yet.
2. The trait-based options, which _could_ be supported if the trait were
enabled currently error if the trait is not enabled. It's likely that
we'll want to treat this similarly to (1)—that is, we'll log an error
diagnostic and continue, or we could throw an actual error.