Rework protobuf codegen to support OTLP/HTTP exporter (#213)
## Motivation

As part of the roadmap for 1.0 (see #206), we intend to provide an
OTLP/HTTP exporter (#205). This exporter will support two modes:
`OTLP/HTTP+protobuf` and `OTLP/HTTP+json`. Both of these make use of
both the OTLP Protobuf types and the gRPC message types in their body,
and differ only in serialization format: binary protobuf, and
JSON-encoded protobuf, respectively.

To support this we need the `OTLPHTTP` target to have access to the
generated Swift.

## Modifications

- Add `make dump` target to help developing Makefile
- Use consistent `subst` Makefile rule for `OTLPCore` and `OTLPGRPC`
outputs
- Regenerate and commit Swift code with latest protoc plugins (but with
existing layout)
- Update Makefile to generate gRPC service definitions in `OTLPCore` too
- Regenerate and commit Swift code with latest protoc plugins (now with
new layout)

## Result

The types needed to implement the OTLP/HTTP exporter are now in place
and available from the `OTLPHTTP` module.