Add stubs for the OTLP/HTTP exporter types (#214)
## Motivation

As part of the roadmap for 1.0 (see #206), we intend to provide a
OTLP/HTTP exporters (#205). We'll need to support independent exporter
config for each of the exporters and each will naturally use different
types in its HTTP API. However, we'd like to structure these types to
reuse as much common logic as practical.

## Modifications

- Factor out OTLP conversion logic from GRPC exporter into OTLPCore
- Factor out constant instrumentation scope to a static stored property
- Stub out the OTLP/HTTP exporters

## Result

The types needed to implement the OTLP/HTTP exporter are now in place
and available from the `OTLPHTTP` module.

## Notes

This is just getting the shape of the types right, which will unblock
some more parallel work to:
- Implement the common exporter logic
- Implement the logging exporter