Have OTelPeriodicExportingMetricsReader read and export on graceful shutdown (#216)
## Motivation

During the implementation of the OTLP/HTTP exporters, I discovered a bug
with the `OTelPeriodicExportingMetricsReader`: during graceful shutdown
it did not perform a final read from the producer and export, as
required by the spec[^1]. The desired behaviour is similar to what
exists in the `BatchSpanProcessor`, which _does_ do this on graceful
shutdown, and for which there is a test.

## Modifications

- Have `OTelPeriodicExportingMetricsReader` read and export on graceful
shutdown
- Add a test that exercises graceful shutdown and checks for the
expected number of `Producer.produce`, `Exporter.forceFlush`, and
`Exporter.shutdown` calls

## Result

`OTelPeriodicExportingMetricsReader` performs a final read and export on
graceful shutdown.

[^1]: https://opentelemetry.io/docs/specs/otel/metrics/sdk/