Compute OTLP/HTTP endpoint from shared key with signal-specific path (#217)
## Motivation

As part of the roadmap for 1.0 (see #206), we're providing OTLP/HTTP
exporters (#205). The OTel spec has specific guidance on how the
OTLP/HTTP endpoints should be computed from a combination of the shared
and signal-specific keys[^1].

## Modifications

- Potentially add signal-specific endpoint path to endpoint during
configuration resolution.

## Result

`OTel.Configuration.OTLPExporterConfiguration.endpiont` is computed
differently when `.protocol` is `.httpProtobuf` or `.httpJSON` vs.
`.grpc`, according to the OTel spec.

[^1]:
https://opentelemetry.io/docs/specs/otel/protocol/exporter/#endpoint-urls-for-otlphttp