Implement OTel.bootstrap API in terms of backend APIs (#219)
## Motivation

As part of the roadmap for 1.0 (see #206), we're providing a simplified
bootstrap API (#205). This is composed on top of the configuration and
make backend APIs, all of which were stubbed out in #210. This PR
implements the main, front-door API, `OTel.bootstrap(configuration:)`
API in terms of the other APIs.

## Modifications

- Implement bootstrap APIs in terms of backend APIs

## Result

The bootstrap API is fully implemented, although it it still based on
the backend APIs, which remain unimplemented.

## Notes

- The PRs to create the backends will follow.
- The PR to add API docs to this function will follow (since they will
likely have more bike-shedding).