Ensure HTTP client is shutdown in OTLPHTTPExporter.deinit (#223)
## Motivation

In #218 we added the implementation for the OTLP/HTTP exporters. As part
of the review we discussed the options for shutting down the HTTP client
and opted to not use `deinit` and to provide a sync shutdown method that
folks were expected to call in a defer block when creating an exporter.
This turns out to not be adequate for all scenarios.

The usual flow for the OTLP/HTTP exporter is that its shutdown is
handled by the object that holds the exporter. In some narrow scenarios,
this type will have been created but the holding type not successfully
created or started, e.g. when there is a misconfiguraiton. In these
scenarios, the user should be presented with a clear error that they can
debug. Having the application crash because HTTP client inside the
exporter was not shutdown will worsen the experience.

## Modifications

- Shutdown the HTTP client in `OTLPHTTPExporter.deinit`.

## Results

More robust.