Add API docs to OTel.bootstrap (#222)
## Motivation

As part of the roadmap for 1.0 (see #206), we're providing a simplified
bootstrap API (#205). This was implemented in #219 but the docs were
deliberately deferred so we could take our time to get them polished,
since this is the primary entry point for the library. This PR adds the
API documentation to `OTel.bootstrap(configuration:)`.

## Modifications

- Add API docs to `OTel.bootstrap(configuration:)`

## Result

Good documentation available to adopters for the primary API of the
library.