Add API docs to OTel.Configuration (#215)
## Motivation

As part of the roadmap for 1.0 (see #206), we added the configuration
API from the proposal (#205). This was added #210 but deliberately
deferred adding the documentation, which would likely attract more
discussion during review. This PR adds the API docs for the
configuration API types.

## Modifications

- Add API docs to `OTel.Configuration` and friends.

## Result

The API types nested in `OTel.Configuration` have API documentation.

## Notes

It would be nice to make use of Snippets for code examples used in DocC
documentation to CI them. There's an item on the roadmap to investigate
this.

---------

Co-authored-by: Moritz Lang <16192401+slashmo@users.noreply.github.com>