Use file-based TLS config in OTLPGRPCMockCollector (#224)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning on refactoring
the OTLP/gRPC exporters. This is a stepping stone that updates the mock
collector to use file-backed TLS config, which will be useful when we
move over the exporters to be configured using `OTel.Configuration`.

## Modifications

- Use file-based TLS config in OTLPGRPCMockCollector

## Result

All tests still remain, but unblocks moving the OTLP/gRPC exporters to
use `OTel.Configuration` in a subsequent PR.