Factor out common parts of OTLP/gRPC exporters (#225)
## Motivation

As part of the roadmap for 1.0 (see #206), we're reworking the OTLP/gRPC
exporters in a number of ways, including supporting mTLS, updating to
gRPC Swift v2, and adding a logging exporter. To support this effort,
we'd first like to unify the underlying exporter logic used by the
OTLP/gRPC metrics and tracing exporters, which is currently duplicative.

## Modifications

- Factor out common parts of OTLP/gRPC exporters

## Result

We now have a single, shared place for the OTLP/gRPC exporter core
logic, which will make it simpler to iterate on and extend to support
logging.

## Notes

- This does not attempt to make any improvements to the logic, simply to
move it into a shared place as an incremental step.