Simplify Makefile by building instead of downloading protoc plugins (#232)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning to update to
gRPC Swift v2. We'd like to make incremental progress toward the upgrade
and this initial PR simplifies the code generation with the existing
gRPC Swift 1.x.

The current codegen is driven by a Makefile, which goes to some lengths
to get the required protoc plugins: it has a hardcoded version and
checksum for the binary artifacts that it downloads from the associated
GitHub release page and unpacks them on a path, with caching—I think I
added all this, so sorry about that! It turns out that SwiftPM can build
executable products of a dependency so the protoc plugins can simply be
built using e.g. `swift build --product protoc-gen-swift`, which allows
for the simplification of the Makefile.

## Modifications

- Simplify Makefile by building instead of downloading protoc plugins
- Rerun code generation (separate commit in PR) — minor, non-functional
changes because of newer 1.x plugin.
 
## Result

- Simpler Makefile: easier to maintain.
- Incremental progress toward gRPC Swift v2 upgrade.