Initialize OTLP/gRPC exporters from OTel.Configuration and add mTLS support (#228)
## Motivation

As part of the roadmap for 1.0 (see #206), we're reworking the OTLP/gRPC
exporters in a number of ways, including supporting mTLS, updating to
gRPC Swift v2, and adding a logging exporter. Following on from #225,
this PR moves to using `OTel.Configuration` for the initializers and
also brings mTLS support.

## Modifications

- Initialize OTLP/gRPC exporters from OTel.Configuration and add mTLS
support

## Result

OTLP/gRPC exporters are now initialized internally from
`OTel.Configuration` and gain support for mTLS.