Add configuration API for the batch log record processor (#229)
## Motivation

As part of the roadmap for 1.0 (see #206), we're adding support for a
logging backend. The `OTelBatchLogRecordProcessor` was already
implemented some time back, but we held off adding the exporter, pending
the expected work for the 1.0 overhaul.

This PR is the first that starts to consolidate the logging work and
provide a functioning backend. It adds the additional configuration
datamodel needed to configure the batch log record processor according
to the OTel spec.

## Modifications

- Add configuration API for the batch log record processor
 
## Result

Configuration for the logging backend can now be expressed as part of
the `OTel.Configuration` object.