Update SwiftFormat rules to not indent conditional compilation blocks (#233)
## Motivation

We have more code that requires conditional compilation blocks with both
the use of traits, and the use of Swift 6.1 features when available. The
current SwiftFormat rules indent the conditional compilation blocks
which is a bit annoying when it's the whole file. Additionally it
creates future diff churn when then they are changed because all the
code within them will need outdenting.

## Modifications

- Update swiftformat rules to not indent conditional compilation blocks.
- Reformat with the new rules.
- Remove the special comments where this was being disabled on an ad-hoc
basis.

## Result

Easier to edit and read with conditional compilation.