Add tests to check bootstrap API does bootstrap subsystems (#226)
## Motivation

As part of the roadmap for 1.0 (see #206), we're adding new APIs to
bootstrap the Swift observability APIs (#205). These have been
implemented in previous PRs, but it would be nice to test the side
effect of the bootstrap.

Historically this has been challenging, since bootstrapping these
subsystems is a one-time, process-global operation, and attempting a
second bootstrap would result in a fatal error from the underlying Swift
observability API package.

Thankfully, Swift Testing now has exit-tests, which can be used to run a
test (or part of a test) in its own process. A primary motivation for
that feature is testing for fatal errors, but it can also be used to
validate that the forked process exited with success, which makes it a
great fit for testing the side-effects of calling the `OTel.bootstrap`
API: specifically, that it has indeed bootstrapped the underlying
subsystem. Additionally, we'd can use a similar technique to confirm
that the `make[Logging|Metrics|Tracing]Backend` APIs have _not_ resulted
in a bootstrap.

## Modifications

- Add tests for bootstrap APIs

## Result

- Test coverage that `OTel.bootstrap` _has_ indeed caused a bootstrap
- Test coverage that `OTel.make[Metrics|Tracing]Backend` has _not_
caused a bootstrap