Implement the OTLP/HTTP log exporter (#230)
## Motivation

As part of the roadmap for 1.0 (see #206), we're adding support for a
logging backend. The `OTelBatchLogRecordProcessor` was already
implemented some time back, but we held off adding the exporter, pending
the expected work for the 1.0 overhaul.

This PR adds the OTLP/HTTP logging exporter.

## Modifications

- Implement the OTLP/HTTP log exporter
 
## Result

OTLP/HTTP exporter exists for logging backend.