Use scoped static methods for mock OTLP/gRPC collector in tests (#235)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning to update to
gRPC Swift v2. We'd like to make incremental progress toward the upgrade
and this initial PR simplifies the code generation with the existing
gRPC Swift 1.x.

As we move to both gRPC v2 and add logging support we'll need to make
use of the mock collector some more, including from end-to-end tests.

This PR moves the mock collector to provide scoped static methods for
testing.

## Modifications

- Use scoped methods for OTLPGRPCMockCollector
 
## Result

Mock collector used in tests is easier to reason about.