Share generated protobuf messages between HTTP and gRPC exporters (#234)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning to update to
gRPC Swift v2. We'd like to make incremental progress toward the upgrade
and this PR further simplifies the code generation with the existing
gRPC Swift 1.x.

When implementing the OTLP/HTTP exporter, I couldn't work out how to
share the generated messages between the targets so we generated them
twice. Turns out I was holding the plugin wrong and this is possible.

## Modifications

- Update Makefile to remove the duplicated message types
- Rerun code generation (separate commit in PR)
 
## Result

The message types in `OTLPCore` are now shared between `OTLPHTTP` and
`OTLPGRPC`.