Add run method requirement to exporter protocols (#236)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning to update to
gRPC Swift v2 and we'd like to make incremental progress toward the
migration.

There are a number of internal types that conform to `Service` and need
to be run in order to perform their background work, e.g.
`BatchSpanProcessor`, `OTelPeriodicExportingMetricReader`, etc. However,
the current exporter protocols, `OTelMetricExporter`,
`OTelSpanExporter`, and `OTelLogRecordExporter` do not have a run method
because, until now they haven't needed one, and are just called by the
type that holds them.

With gRPC Swift v2, the client connection now has a run method which
takes over the task and so this will need to become part of the startup
flow.

This PR adds that requirement, to make the upcoming v2 migration PR
simpler.

## Modifications

- Add run method requirement to span and metric exporter protocols
- Call the exporter run method from types that hold an exporter
- Add no-op run methods for all concrete exporter types that don't have
background work
- Update tests to use exporter run method
 
## Result

All exporters now conform to `Service` and have a place to perform
background work.