Implement OTLP/gRPC log record exporter (#238)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning to update to
gRPC Swift v2 and add a log record exporter. These things are
intertwined and we've been making incremental progress towards both in
parallel in recent PRs.

Before moving the OTLP/gRPC exporters over to gRPC Swift v2, it would be
good to get the log record exporter in place so that it comes along for
the v2 upgrade ride.

This PR adds the log record OTLP/gRPC exporter, still using gRPC Swift
v1.

## Modifications

- Implement OTLP/gRPC log record exporter
 
## Result

We have a OTLP/gRPC log record exporter