Add documentation to the backend APIs (#241)
## Motivation

As part of the roadmap for 1.0 (see #206), we're providing APIs to make
backends for the Swift observability APIs (#205). The APIs were added
already according to the accepted API proposal, but the documentation
was deferred since it would likely attract more feedback during review.

This PR adds documentation to the backend APIs: `makeLoggingBackend`,
`makeMetricsBackend`, and `makeTracingBackend`.

## Modifications

- Add documentation to the backend APIs

## Result

API documentation added to `makeLoggingBackend`, `makeMetricsBackend`,
and `makeTracingBackend`.