Implement OTel.makeLoggingBackend and add to bootstrap and end-to-end tests (#242)
## Motivation

As part of the roadmap for 1.0 (see #206), we're adding support for a
logging backend and have been making incremental progress to it in
recent PRs.

Now that we have both the OTLP/HTTP and OTLP/gRPC exporters added (in
#230 and #238, respectively), we can now implement the API to make the
logging backend.

This PR adds the implementation for `makeLoggingBackend`.

## Modifications

- Implement OTel.makeLoggingBackend
- Extend bootstrap tests with tests for logging backend
- Extend end-to-end tests with tests for logging backend
 
## Result

Adopters can now bootstrap the logging backend through the `bootstrap()`
and `makeLoggingBackend()` APIs.

## Note

There's an expected failure right now in the end-to-end tests, which
check that the expected number of log events are emitted to the
collector. This is failing because we have yet to move the internal
diagnostic logging to a dedicated logger, so that it also going out the
exporter.