Rename diagnostic logging configuration API for clarity (#244)
## Motivation

As part of the roadmap for 1.0 (see #206), we're overhauling the public
API and want to make it as obvious as possible to use. One source of
confusion is the configuration for the internal diagnostic logger, which
is currently configured using `OTel.Configuration.logger` and
`OTel.Configuration.logLevel`. These influence the diagnostic logging of
Swift OTel itself, and _not_ the logging backend provided by Swift OTel
for use with Swift Log.

This PR updates the API to rename the configuration for the diagnostic
logger to avoid this confusion.

## Modifications

- Rename diagnostic logging configuration API for clarity 

## Result

The following APIs have been renamed:

- `OTel.Configuration.logLevel` ->
`OTel.Configuration.diagnosticLogLevel`
- `OTel.Configuration.logger` -> `OTel.Configuration.diagnosticLogger`
- `OTel.Configuration.LoggerSelection` ->
`OTel.Configuration.DiagnosticLoggerSelection`