Add API to configure the default logging level of the logging backend (#243)
## Motivation

As part of the roadmap for 1.0 (see #206), we're adding support for a
logging backend. The API to make and bootstrap the backend landed in
#242 but we didn't have any way to configure the logging level of
loggers returned by the factory by default.

This PR adds API for configuring that logging level.

## Modifications

- Add API to configure the default logging level of the logging backend
- Add `trace` to the configuration options
 
## Result

Configuration for the logging backend now includes the default log level
applied to loggers returned from the factory.

## Notes

We are only supporting a subset of what can typically be expressed by
Swift Log, since these are the only ones that overlap with the OTel
spec[^1].

[^1]:
https://opentelemetry.io/docs/specs/otel/logs/data-model/#field-severitynumber