Update OTLP/gRPC exporters to gRPC Swift v2 (#245)
## Motivation

As part of the roadmap for 1.0 (see #206), we're planning to update to
gRPC Swift v2. We've been making incremental progress toward that goal
in a number of previous PRs. This PR is the final step and, hopefully,
because of the incremental steps, should be reasonably straightforward
to review.

## Modifications

- Update dependencies for gRPC Swift v2
- Makefile updates for grpc-swift-2
- Rerun code generation
- Update OTLP/gRPC exporters to gRPC Swift v2

## Result

OTLP/gRPC exporters are now using gRPC Swift v2.

## Notes

The commits have been separated so that they can be reviewed in
isolation and to avoid having to review the result of running the code
generation.

The diffstat seems large, but when excluding the generated code, it is
quite modest:

```console
% git diff --stat upstream/main | tail -1
 18 files changed, 1451 insertions(+), 1102 deletions(-)

% git diff --stat upstream/main -- ":(exclude)**/Generated/**" | tail -1
 9 files changed, 206 insertions(+), 190 deletions(-)
```