Move diagnostic logging to dedicated, config-based diagnostic logger (#247)
## Motivation

As part of the roadmap for 1.0 (see #206) we're adding support for a
logging backend, which we've been making incremental progress toward.
The backend was implemented in #242 but the tests have been marked as
expecting a failure because all the diagnostic logging, internal to
Swift OTel itself, is currently using loggers provided by Swift Log, and
therefore are also included in the OTLP logging export, polluting user
logs.

The spec says that diagnostic logging from an OTel library should be
logged in some other way, using a language-specific convention[^1] and
our accepted API for 1.0 (see #205) has configuration for a the
diagnostic logger, which was recently renamed in #244.

We should be using this configuration to construct a logger that _does
not_ use the bootstrapped log factory and pass this to every type in the
Swift OTel implementation that has its own logger.

## Modifications

- Add explicit logger param to all types that use a logger
- Add overloads with logging disabled in tests
- Add internal OTel.Configuration.makeDiagnosticLogger()
- Use config-based diagnostic logger in public APIs
- Update tests that explicitly check for diagnostic logging

## Result

Internal logging by Swift OTel uses the configured diagnostic logger,
instead of the bootstrapped logging factory.

# Notes

This PR has been split into incremental commits to help with review.

[^1]:
"https://opentelemetry.io/docs/specs/otel/error-handling/#self-diagnostics