Add tests for configuration overrides (with some fixes) (#248)
## Motivation

As part of the roadmap for 1.0 (see #206) we're providing config-driven
bootstrap and backend APIs, which support environment variable overrides
defined in the OTel spec (see #205).

This PR adds tests for the logic that applies the environment variable
overrides.

## Modifications

- Add tests for configuration environment overrides
- Fix config key: `OTEL_PROPARGATORS` -> `OTEL_PROPAGATORS`
- Fix config key: `OTEL_BSP_EXPORT_BATCH_SIZE` ->
`OTEL_BSP_MAX_EXPORT_BATCH_SIZE`
- Fix config key: `OTEL_BLRP_EXPORT_BATCH_SIZE` ->
`OTEL_BLRP_MAX_EXPORT_BATCH_SIZE`
- Support surrounding whitespace in parsed headers
 
## Result

- Configuration overrides are well tested
- `OTEL_PROPAGATORS` is now handled properly
- `OTEL_BSP_MAX_EXPORT_BATCH_SIZE` is now handled properly
- `OTEL_BLRP_MAX_EXPORT_BATCH_SIZE` is now handled properly
- `OTEL_EXPORTER_OTLP_HEADERS` now supports optional whitespace as per
the W3C spec

## Notes

There are a couple of things that currently do _not_ work as we want
them to and there are variables that we should support but don't yet.
These are called out in the tests, either with a `.disabled(comment:)`
Testing trait, or with a `recordIssue`.