Use enum-backed wrappers to remove existentials (#250)
## Motivation

As part of the roadmap for 1.0 (see #206), we're no longer exposing the
internal types as public API (#205) and instead using opaque return
types for both the factories that the users should bootstrap and the
service the user should run.

However, the backing implementation for much of this library is
comprised of layered generic types, for example:

```swift
OTelTracer<
  OTelRandomIDGenerator<SystemRandomNumberGenerator>,
  OTelTraceIDRatioBasedSampler,
  OTelW3CPropagator,
  OTelBatchSpanProcessor<OTLPHTTPSpanExporter, ContinuousClock>,
  ContinuousClock
>
```

When returning opaque types, the type must still be known at compile
time and be the same type on all branches.

This can be achieved in two ways:

1. Return a concrete wrapper type that holds an existential.
2. Return a concrete wrapper type that is an enum.

(1) is a poor trade for APIs that return only a fixed set of types since
we introduce an existential, which may have performance implications.

(2) is a better choice for a closed set of types since it introduces
minimal overhead.

## Modifications

- Refactor the implementation of the backend APIs to use enum-backed
wrapper types for the closed set of generic types.
- Remove the existential on the API surface.

## Result

- Cleaner implementation for backend APIs
- Removed one layer of existential, which will result in better
performance
- `makeMetricsBackend` now returns a `some MetricsFactory`
- `makeTracingBackend` now returns a `some Tracer`