Add support for configuring the traces sampler (#251)
## Motivation

As part of the roadmap for 1.0 (see #206), we're using a config-based
public API (#205). Until now we've been hardcoding the sampler used for
tracing to `.alwaysOn`, but this should be configurable, both in-code
and using the `OTEL_TRACES_SAMPLER` and `OTEL_TRACES_SAMPLER_ARG`
environment variables.

## Modifications

- Add configuration types for sampler selection
- Add enum-backed wrapper for sampler for use as generic argument
- Add environment configuration overrides for sampler
- Add tests for sampler configuration environment overrides
- Use configured sampler in `makeTracingBackend(configuration:)`

## Result

Users can now configure the sampler used for traces, both in-code, and
using `OTEL_TRACES_SAMPLER`.