Populate service.name resource attribute on export (#254)
## Motivation

During export, the resource attribute `service.name` is expected to be
populated. It's value should be taken from the explicit service name
configuration value, if set, falling back to the `service.name` resource
attribute, if set, falling back to `unknown_service`.

> If service.name is also provided in OTEL_RESOURCE_ATTRIBUTES, then
OTEL_SERVICE_NAME takes precedence.
> — source:
https://opentelemetry.io/docs/languages/sdk-configuration/general/#otel_service_name


The logic to handle the `OTEL_SERVICE_NAME` and
`OTEL_RESOURCE_ATTRIBUTES` is already correct, but `service.name` was
not appearing in the resource attributes when
`OTel.Configuration.serviceName` was set.

This was because `OTelResource.init(configuration:)`, which is then
passed to the exporters, was only taking the values in the resource
attributes.

## Modifications

- Populate service.name resource attribute on export

## Result

The `service.name` attribute is now set on exports and honors the
`serviceName` configuration value and `OTEL_SERVICE_NAME` environment
variable.