Add Logging metadata provider API with end-to-end tests (#256)
## Motivation

As part of the 1.0 release, we wanted to include a
`Logger.MetadataProvider` for users who are not using the OTLP logging
backend, but wish to have their logs correlated with the spans produced
by the tracing backend.

There was also some discussion about whether we should be adding this
logging metadata provider to the OTLP backend itself. This turns out to
be unnecessary, since the OTLP log records already contain the span and
trace IDs.

## Modifications

- Add configuration datamodel for logging metadata provider
- Add `OTel.makeLoggingMetadataProvider` API
- Add end-to-end test to check span and trace IDs in OTLP log records
- Add end-to-end test using OTel metadata provider with StreamLogHandler

## Result

New API `OTel.makeLoggingMetadataProvider`