Correctly thread through the BLRP configuration (#257)
## Motivation

In `OTel.makeLoggingBackend` we construct the chain of objects used for
the OTLP logging backend. At the time, we didn't have the
`BatchLogRecordProcessorConfiguration` shim for the new configuration
datamodel so we just used the default config, always. Now that the
configuration API supports configuring the BLRP, we should be using
that.

## Modifications

- Correctly thread through the BLRP configuration

## Result

Batch log record processor configuration is actually used during
bootstrap.