Relax package requirements to macOS 13, iOS 16, tvOS 16, watchOS 9, visionOS 1 (#261)
## Motivation

When we updated the OTLP/gRPC exporters to gRPC Swift v2, we temporarily
raised the minimum platform requirements for Swift OTel to match the
requirements of gRPC Swift v2. These are pretty new (iOS 18 aligned),
but the rest of the package only depends on iOS 16 aligned APIs (e.g.
`Clock`, `Duration`, etc.).

Prior to this they were set to macOS 13 (incompletely: only macOS was
specified).

There's a continuum of options here, including removing the package
platform requirements entirely and using availability annotations on the
entire API surface. Since the majority of the API depends on `Clock`,
`Duration`, and friends, we have a practical minimum of macOS 13, iOS
16, etc., and the package would likely offer an empty API surface on
platforms older than these.

This approach would be necessary if a _library_ package wanted to take a
_new_ dependency on Swift OTel, that currently has more relaxed platform
requirements (including no stated requirements), without resulting in a
SemVer major for that package.

However, Swift OTel is intended for leaf/application packages as its
principal offering is APIs for bootstrapping the Swift observability
backends.

During API review on the forums, we did have a concrete request to not
_raise_ the platform requirements from what they were in the 0.x
releases. This was justified by app developers who need to support a few
years of OS releases back, which seemed reasonable.

At this point, an important note is that it's always possible to relax
the requirements of this package without a SemVer major, but increasing
them is a breaking change.

The decision at the contributors meetup was to proceed with a pragmatic
middle-ground: to lower the package platform requirements to macOS 13,
iOS 16, etc., but defer removing them entirely for now.

## Modifications

- Define PlatformRequirements helper struct in Package.swift
- Drop package platform requirements and use `@available(gRPCSwift, *)`
where needed

## Result

- Package requirements now macOS 13, iOS 16, tvOS 16, watchOS 9,
visionOS 1
- OTLP/gRPC APIs marked as only available on macOS 15, iOS 18, tvOS 18,
watchOS 11, visionOS 2